jwt working

+++ api/resolvers/User.Resolvers.js
@@ -1,5 +1,10 @@
-const { PubSub } = require("apollo-server-express");
+const { PubSub, AuthenticationError, ForbiddenError } = require("apollo-server-express");
 const pubsub = new PubSub();
+
+const bcrypt = require("bcrypt");
+const saltRounds = 10;
+
+const jwt = require("jsonwebtoken");
 
 const eventName = {
   USER_CREATED: "USER_CREATED",
@@ -17,38 +22,46 @@ module.exports = {
     }
   },
   Mutation: {
-    createUser: (
-      parent,
-      { first_name, last_name, email },
-      { models: { User } },
-      info
-    ) => {
-      return User.create({ first_name, last_name, email }).then(user => {
-        return pubsub
-          .publish(eventName.USER_CREATED, { userCreated: user })
-          .then(done => {
+    createUser: (parent, { first_name, last_name, email, password }, { requester, models: { User } }, info) => {
+      return bcrypt
+        .hash(password, saltRounds)
+        .then(hash => {
+          return User.create({
+            first_name,
+            last_name,
+            email,
+            password: hash
+          });
+        })
+        .then(user => {
+          return pubsub.publish(eventName.USER_CREATED, { userCreated: user }).then(done => {
             return user;
           });
+        });
+    },
+    updateUser(parent, { id, ...patch }, { requester, models: { User } }, info) {
+      if (!requester || requester._id != id) throw new ForbiddenError("Not allowed");
+      return User.findOneAndUpdate({ _id: id }, patch, { new: true }).then(user => {
+        return pubsub.publish(eventName.USER_UPDATED, { userUpdated: user }).then(done => {
+          return user;
+        });
       });
     },
-    updateUser(parent, { id, ...patch }, { models: { User } }, info) {
-      return User.findOneAndUpdate({ _id: id }, patch, { new: true }).then(
-        user => {
-          return pubsub
-            .publish(eventName.USER_UPDATED, { userUpdated: user })
-            .then(done => {
-              return user;
-            });
-        }
-      );
-    },
-    deleteUser: (parent, { id }, { models: { User } }, info) => {
+    deleteUser: (parent, { id }, { requester, models: { User } }, info) => {
+      if (!requester || requester._id != id) throw new ForbiddenError("Not allowed");
       return User.findOneAndDelete({ _id: id }).then(user => {
-        return pubsub
-          .publish(eventName.USER_DELETED, { userDeleted: user })
-          .then(done => {
-            return user;
-          });
+        return pubsub.publish(eventName.USER_DELETED, { userDeleted: user }).then(done => {
+          return user;
+        });
+      });
+    },
+    login(parent, { email, password }, { models: { User } }, info) {
+      return User.findOne({ email }).then(user => {
+        if (!user) throw new AuthenticationError("Bad credentials");
+        return bcrypt.compare(password, user.password).then(hash => {
+          if (!hash) throw new AuthenticationError("Bad credentials");
+          return jwt.sign({ _id: user._id }, process.env.JWTAUTH_KEY);
+        });
       });
     }
   },
